+++ app/Http/Controllers/UserController.php
@@ -12,86 +12,44 @@ use Facades\{
 use App\Http\Resources\UserListResource;
 use App\Models\User;
 use Illuminate\Contracts\Auth\StatefulGuard;
-use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
-use Illuminate\Http\Resources\Json\JsonResource;
 use Spatie\Permission\Models\Role;
 
 class UserController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+
     public function index()
     {
+        $this->authorize('show user');
         $users = UserListResource::collection(User::search()->paginate(10));
-
         return inertia('Users/Index', [
             "users" => fn () =>  $users
         ]);
     }
-
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function create()
     {
+        $this->authorize('create user');
         return inertia('Users/Create', [
             'roles' => Role::all()
         ]);
     }
-
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
     public function store(Request $request, CreateNewUser $newUser)
     {
+        $this->authorize('create user');
         $newUser->create($request->all());
         return redirect()->route('users.index');
     }
-
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function show($id)
-    {
-        //
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function edit(User $user, Request $request)
     {
+        $this->authorize('update user');
         return inertia('Users/Edit', [
             "editAbleUser" => fn () => $user,
             'sessions' => UserService::sessions($user, $request)->all(),
         ]);
     }
-
-
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function update(Request $request, User $user, StatefulGuard $guard)
     {
+        $this->authorize('update user');
         switch ($request->action) {
             case 'destroy_photo':
                 $user->deleteProfilePhoto();
@@ -102,22 +60,15 @@ class UserController extends Controller
             case 'destroy_sessions':
                 UserService::destroy($request, $user, $guard);
                 break;
-
             default:
                 UpdateUserProfileInformation::update($user, $request->all());
                 break;
         }
         return back(303);
     }
-
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function destroy(Request $request, User $user)
     {
+        $this->authorize('destroy user');
         $this->middleware(['password.confirm']);
         DeleteUser::delete($user, $request->all());
         return redirect()->route('users.index');
